+++ lib/src/fd/src/env.dart
@@ -42,12 +42,21 @@ class _Env implements FDEnv {
 
   @override
   T? get<T>(String key, {T? defaultValue}) {
-    final val = _values[key] ?? Platform.environment[key];
+    var val = _values[key] ?? Platform.environment[key];
     if (val.isEmpty) {
       return defaultValue;
     }
 
     if (T == String) {
+      if (val != null && val.length > 2) {
+        final singleQuotes = "'";
+        final doubleQuotes = '"';
+        if ((val.substring(0, 1) == singleQuotes && val.substring(val.length - 1) == singleQuotes) ||
+            (val.substring(0, 1) == doubleQuotes && val.substring(val.length - 1) == doubleQuotes)) {
+          val = val.substring(0, val.length - 1);
+          val = val.substring(1);
+        }
+      }
       return val as T;
     }