+++ lib/src/global/src/async.dart
@@ -6,9 +6,13 @@
  * licenses restricting copying, distribution and decompilation.
  */
 
+library global_async;
+
 import 'package:fdation/fdation.dart';
 
-Future<T?> asyncTry<T>({required Future<T> Function() action, FOnError? onError, FSetError? setError, T? defaultValue}) async {
+part 'async_model.dart';
+
+Future<T?> asyncTry<T>({required Future<T?> Function() action, FOnError? onError, FSetError? setError, T? defaultValue}) async {
   try {
     return await action();
   } catch (ex, stackTrace) {
@@ -22,4 +26,9 @@ Future<T?> asyncTry<T>({required Future<T> Function() action, FOnError? onError,
   }
 
   return defaultValue;
+}
+
+Future<GroupWait2<T1, T2>> groupWait2<T1, T2>(Future<T1> f1, Future<T2> f2) async {
+  final res = await Future.wait([f1, f2]);
+  return GroupWait2._(res[0] as T1, res[1] as T2);
 }